[WIP] - Add meaningful OAuth2 integration.
This change adds a meaningful OAuth2 integration that uses Google API as
an example. The template is usable with other integrations out of the
box that work with authlib's loginpass library.

There are still some FIXME to address, and a few other flows (like
signup) that I need to add. I need to also associate the User object
with the OAuth2Token that they may be associated with.